Reorganized light curve modeling code into new package
