+++ Project.toml
@@ -16,4 +16,6 @@ Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+attitudeFunctions = "7493c2ef-4a11-4f18-aeab-b8ca9ae8f56a"
+lightCurveModeling = "bbb4d786-ea76-4613-b8bd-e0066326f386"
 myFilters = "81456be0-7a9e-4877-8423-9bcd083e1f21"
